refactorings mostly
