+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/building/N4JSBuilderParticipant.java
@@ -31,11 +31,11 @@ import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.SubMonitor;
 import org.eclipse.emf.ecore.resource.ResourceSet;
 import org.eclipse.emf.ecore.util.EcoreUtil;
+import org.eclipse.n4js.external.ExternalProject;
 import org.eclipse.n4js.generator.ICompositeGenerator;
 import org.eclipse.n4js.ui.building.instructions.BuildInstruction;
 import org.eclipse.n4js.ui.building.instructions.CleanInstruction;
 import org.eclipse.n4js.ui.building.instructions.IBuildParticipantInstruction;
-import org.eclipse.n4js.utils.resources.ExternalProject;
 import org.eclipse.xtext.builder.BuilderParticipant;
 import org.eclipse.xtext.builder.EclipseResourceFileSystemAccess2;
 import org.eclipse.xtext.builder.IXtextBuilderParticipant;
