+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/internal/N4JSProjectDependencyStrategy.java
@@ -18,10 +18,10 @@ import java.util.Collection;
 import java.util.List;
 
 import org.eclipse.core.resources.IProject;
+import org.eclipse.n4js.external.ExternalProject;
 import org.eclipse.n4js.projectModel.IN4JSProject;
 import org.eclipse.n4js.ui.projectModel.IN4JSEclipseCore;
 import org.eclipse.n4js.ui.projectModel.IN4JSEclipseProject;
-import org.eclipse.n4js.utils.resources.ExternalProject;
 
 import com.google.inject.Inject;
