+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/projectModel/dependencies/ProjectDependenciesHelper.java
@@ -12,24 +12,35 @@ package org.eclipse.n4js.projectModel.dependencies;
 
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.StringJoiner;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import javax.inject.Inject;
 
 import org.apache.log4j.Logger;
 import org.eclipse.emf.common.util.URI;
+import org.eclipse.n4js.external.ExternalLibraryWorkspace;
+import org.eclipse.n4js.external.NpmLogger;
 import org.eclipse.n4js.internal.N4JSModel;
 import org.eclipse.n4js.internal.N4JSProject;
 import org.eclipse.n4js.projectDescription.ProjectDescription;
 import org.eclipse.n4js.projectDescription.ProjectReference;
 import org.eclipse.n4js.projectModel.IN4JSCore;
 import org.eclipse.n4js.projectModel.IN4JSProject;
+import org.eclipse.n4js.semver.SemverMatcher;
 import org.eclipse.n4js.semver.SemverUtils;
 import org.eclipse.n4js.semver.Semver.NPMVersionRequirement;
+import org.eclipse.n4js.semver.Semver.VersionNumber;
+import org.eclipse.n4js.semver.model.SemverSerializer;
+
+import com.google.common.base.Joiner;
 
 /**
  * Allows to access data in the manifests of the provided {@link IN4JSProject}s. Unlike {@link N4JSModel} will return
@@ -42,6 +53,10 @@ public class ProjectDependenciesHelper {
 	private IN4JSCore core;
 	@Inject
 	private N4JSModel model;
+	@Inject
+	private ExternalLibraryWorkspace externalWS;
+	@Inject
+	private NpmLogger logger;
 
 	/**
 	 * Same as {@link #calculateDependenciesOfProjects(Collection)} but uses all user workspace projects as input.
@@ -92,6 +107,103 @@ public class ProjectDependenciesHelper {
 			messages.add("dependencies to install: ");
 			versionedPackages.forEach((id, v) -> messages.add(" - " + id + v));
 			LOGGER.debug(messages);
+		}
+	}
+
+	/**
+	 * In case some dependencies are shipped code, the following will be done here: All shipped code will be installed
+	 * with the given version and hence shadow the shipped code of the IDE.
+	 * <p>
+	 * In case the dependencies do not contain shipped code, nothing is done.
+	 */
+	public void fixDependenciesToInstall(Map<String, NPMVersionRequirement> dependenciesToInstall) {
+		Map<String, VersionNumber> projectNamesOfShippedCode = externalWS.getProjectInfos();
+		removeDependenciesToShippedCodeIfVersionMatches(dependenciesToInstall, projectNamesOfShippedCode);
+		addDependenciesForRemainingShippedCode(dependenciesToInstall, projectNamesOfShippedCode.keySet());
+		logShippedCodeInstallationStatus(dependenciesToInstall, projectNamesOfShippedCode.keySet());
+	}
+
+	/**
+	 * Removes from map 'dependenciesToInstall' all entries for projects that are in the shipped code, if and only if
+	 * the version provided by the shipped code matches the version requirement.
+	 */
+	private void removeDependenciesToShippedCodeIfVersionMatches(
+			Map<String, NPMVersionRequirement> dependenciesToInstall,
+			Map<String, VersionNumber> projectNamesOfShippedCode) {
+
+		Set<String> prjNames = new HashSet<>(dependenciesToInstall.keySet());
+		for (String projectName : prjNames) {
+			VersionNumber availableVersionInShippedCode = projectNamesOfShippedCode.get(projectName);
+			if (availableVersionInShippedCode != null) {
+				NPMVersionRequirement versionRequirement = dependenciesToInstall.get(projectName);
+				if (versionRequirement != null
+						&& SemverMatcher.matchesStrict(availableVersionInShippedCode, versionRequirement)) {
+					// so remove from list of dependencies to be installed:
+					dependenciesToInstall.remove(projectName);
+				}
+			}
+		}
+	}
+
+	/**
+	 * If map 'dependenciesToInstall' contains at least one project that is among the shipped code projects, this method
+	 * will add new entries to map 'dependenciesToInstall' for all remaining shipped code projects, using the same
+	 * version constraint as for those already in the map.
+	 * <p>
+	 * Rationale is that when shadowing a shipped code project with a newly installed NPM in the library manager, then
+	 * *all* shipped code projects need to be shadowed, because shipped code is now published in clusters in which each
+	 * project depends on the others with a fixed version.
+	 */
+	private void addDependenciesForRemainingShippedCode(Map<String, NPMVersionRequirement> dependenciesToInstall,
+			Set<String> projectNamesOfShippedCode) {
+		Set<String> projectNamesOfShippedCodeToInstall = new HashSet<>(dependenciesToInstall.keySet());
+		projectNamesOfShippedCodeToInstall.retainAll(projectNamesOfShippedCode);
+		if (!projectNamesOfShippedCodeToInstall.isEmpty()
+				&& projectNamesOfShippedCodeToInstall.size() < projectNamesOfShippedCode.size()) {
+			Map<String, NPMVersionRequirement> versionRequirementsOfShippedCodeToInstall = projectNamesOfShippedCodeToInstall
+					.stream()
+					.map(id -> dependenciesToInstall.get(id))
+					.collect(Collectors.toMap(SemverSerializer::serialize, Function.identity(), (vr1, vr2) -> vr1));
+			NPMVersionRequirement versionConstraint = versionRequirementsOfShippedCodeToInstall.values().stream()
+					.findFirst().orElse(null);
+			if (versionConstraint != null) {
+				if (versionRequirementsOfShippedCodeToInstall.size() > 1) {
+					logger.logInfo("WARNING: differing version requirements for shipped code: "
+							+ Joiner.on(", ").join(versionRequirementsOfShippedCodeToInstall.values())
+							+ "; using: " + versionConstraint);
+				}
+				for (String id : projectNamesOfShippedCode) {
+					if (!projectNamesOfShippedCodeToInstall.contains(id)) {
+						dependenciesToInstall.put(id, versionConstraint);
+					}
+				}
+			}
+		}
+	}
+
+	/** Pure logging, won't change the arguments. */
+	private void logShippedCodeInstallationStatus(Map<String, NPMVersionRequirement> dependenciesToInstall,
+			Set<String> projectNamesOfShippedCode) {
+		// compute status of shipped code installation
+		Map<String, NPMVersionRequirement> shippedCodeToInstall = new HashMap<>(dependenciesToInstall);
+		Set<String> nonShadowedShippedCode = new HashSet<>(projectNamesOfShippedCode);
+		shippedCodeToInstall.keySet().retainAll(projectNamesOfShippedCode);
+		nonShadowedShippedCode.removeAll(shippedCodeToInstall.keySet());
+		// report status to user
+		if (shippedCodeToInstall.isEmpty()) {
+			logger.logInfo("Not going to shadow any shipped code with installed packages.");
+		} else {
+			String separator = "\n\t";
+			logger.logInfo("The following installed packages will shadow shipped code:" + separator
+					+ shippedCodeToInstall.entrySet().stream()
+							.map(e -> e.getKey() + "@" + e.getValue())
+							.collect(Collectors.joining(separator)));
+			if (nonShadowedShippedCode.isEmpty()) {
+				logger.logInfo("The entire shipped code will be shadowed.");
+			} else {
+				logger.logInfo("WARNING: shipped code will be shadowed only partially; non-shadowed shipped projects:"
+						+ separator + Joiner.on(separator).join(nonShadowedShippedCode));
+			}
 		}
 	}
